Redirect 'swift --version' stderr to /dev/null
With the new swift driver running 'swift --version' will always print "swift-driver version: X.Y.Z" to stderr before printing the normal version info to stdout.
This causes all of the version checks as swiftenv runs to print the driver over and over.
The other option is to instead pass the flag '-disallow-use-new-driver' but at some point the new driver will most likely become the only driver so this future proofs against that.